requirements.txt
Flask
sqlite3
pillow